requirements.txt
protobuf==4.22.0
grpcio==1.51.3
grpcio-tools==1.51.3
certifi==2022.12.7
minio==7.1.13
urllib3==1.26.14
config==0.5.1
